+++ crates/router/src/core/webhooks/webhook_events.rs
@@ -231,7 +231,7 @@ pub async fn retry_delivery_attempt(
         .change_context(errors::ApiErrorResponse::InternalServerError)
         .attach_printable("Failed to parse webhook event request information")?;
 
-    Box::pin(super::trigger_webhook_and_raise_event(
+    Box::pin(super::outgoing::trigger_webhook_and_raise_event(
         state.clone(),
         business_profile,
         &key_store,
